🎈 feat(mobile/shared/lib): reorganize storage and view imports for better structure
