+++ apps/mobile/src/layers/entities/session/model/account-association-session-token.ts
@@ -1,4 +1,4 @@
-import { atomWithSecureStorage, secureStorageKeys } from "../../../shared/lib";
+import { atomWithSecureStorage, secureStorageKeys } from "../../../shared/lib/storage";
 
 export const accountAssociationSessionTokenAtom = atomWithSecureStorage<string | null>(
 	secureStorageKeys.accountAssociationToken,
