+++ apps/mobile/src/layers/entities/session/model/last-login-method.ts
@@ -1,4 +1,4 @@
-import { atomWithGlobalStorage, globalStorageKeys } from "../../../shared/lib";
+import { atomWithGlobalStorage, globalStorageKeys } from "../../../shared/lib/storage";
 
 export type LoginMethod = "email" | "google" | "discord";
