+++ apps/mobile/src/layers/entities/session/model/session-token.ts
@@ -1,5 +1,5 @@
 import { atom } from "jotai";
-import { atomWithSecureStorage, secureStorageKeys } from "../../../shared/lib";
+import { atomWithSecureStorage, secureStorageKeys } from "../../../shared/lib/storage";
 
 export const sessionTokenAtom = atomWithSecureStorage<string | null>(secureStorageKeys.sessionToken, null);
