+++ apps/mobile/src/layers/shared/lib/data/index.ts
@@ -1,0 +1,3 @@
+export { bytesToHex } from "./converter";
+
+export type { ValueOf } from "./types";
