+++ apps/mobile/src/layers/shared/lib/data/types.ts
