--- apps/mobile/src/layers/shared/lib/index.ts
@@ -1,19 +1,0 @@
-export { useLayoutInsets } from "./view/layout-insets";
-export { vh, vw, screenHeight, screenWidth } from "./view/viewport";
-
-export {
-	createSecureJSONStorage,
-	createUserJSONStorage,
-	createGlobalJSONStorage,
-} from "./storage/json-storage";
-export { secureStorageKeys, globalStorageKeys, userStorageKeys } from "./storage/keys";
-export {
-	atomWithGlobalStorage,
-	atomWithSecureStorage,
-	atomFamilyWithUserStorage,
-} from "./storage/atom";
-export type { JSONStringifyable } from "./storage/types";
-
-export { bytesToHex } from "./converter";
-
-export type { ValueOf } from "./types";
