+++ apps/mobile/src/layers/shared/lib/storage/keys.ts
@@ -7,7 +7,7 @@ export const globalStorageKeys = {
 	theme: "THEME",
 	lastLoginMethod: "LAST_LOGIN_METHOD",
 	user: "USER",
-	visitableSetupPage: "VISITABLE_SETUP_PAGE",
+	visitPersonalizePageFlag: "VISIT_PERSONALIZE_PAGE_FLAG",
 } as const satisfies Record<string, string>;
 
 export const userStorageKeys = {
