--- apps/mobile/src/layers/shared/ui/index.ts
@@ -1,4 +1,0 @@
-export { Wave } from "./wave/wave";
-export { ContinueWithEmailButton } from "./continue-with-method-button/email";
-export { ContinueWithGoogleButton } from "./continue-with-method-button/google";
-export { ContinueWithDiscordButton } from "./continue-with-method-button/discord";
