+++ apps/mobile/src/layers/shared/ui/wave/index.ts
@@ -1,0 +1,1 @@
+export { Wave } from "./wave";
